No more DeviceDriver Interface. Now we just have a single class.
